backend/requirements.txt
# Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# MongoDB
pymongo==4.6.0

python-dotenv==1.0.0 
certifi==2025.6.15